renamed and updated test scripts for rename
