+++ collectionmerge.py
